Refactor contributivity measures as class methods
